refactor(session): estrazione IA in services/ai/ (sprint-010 / issue #2)
Chiude parzialmente l'issue #2 del playtest design backlog con l'estrazione
dell'orchestrazione IA SIS dal router session.js in un modulo dedicato
services/ai/. Prep necessaria per implementare issue #10 (stati emotivi
panic/rage/confused/focused) in modo pulito.

Struttura nuova:

services/ai/policy.js (84 linee, puro)
  - DEFAULT_ATTACK_RANGE (ora autoritativo, importato da session.js)
  - DEFAULT_MAX_HP_FALLBACK, LOW_HP_RETREAT_THRESHOLD
  - manhattanDistance(a, b) — copia locale per evitare dipendenza
    circolare con session.js (che ha la sua utility omonima)
  - stepAway(from, to, gridSize?) — passo di ritirata, null se
    cornered
  - selectAiPolicy(actor, target) — decide { rule, intent } basato
    su HP ratio + distance. Pensata per essere estendibile in futuro
    con pre-processo stati emotivi:
      if (actor.status?.rage) return RAGE attack
      if (actor.status?.panic) return PANIC retreat
      // ... policy normale

services/ai/sistemaTurnRunner.js (217 linee, orchestratore con
dependency injection)
  - createSistemaTurnRunner(deps) factory
  - Ritorna async function runSistemaTurn(session)
  - Inietta: pickLowestHpEnemy, manhattanDistance, stepTowards,
    performAttack, buildAttackEvent, buildMoveEvent, emitKillAndAssists,
    appendEvent, gridSize
  - Mantiene TUTTA la logica del turno SIS: dual-action loop, fallback
    cornered (con flag corneredThisTurn per evitare oscillazioni),
    skip con reason, overlap guard, retreat vs approach distinction
  - Action descriptor output invariato (compat col frontend sprint-009)

routes/session.js (−239, +25)
  - Rimossa definizione inline di stepAway (38 linee)
  - Rimossa definizione inline di selectAiPolicy (16 linee)
  - Rimossa definizione inline di runSistemaTurn (168 linee)
  - Rimosso const DEFAULT_ATTACK_RANGE locale (autoritativo in policy.js)
  - Aggiunti import dal nuovo modulo
  - createSistemaTurnRunner istanziato dentro createSessionRouter con
    dipendenze del closure (performAttack, buildAttackEvent, ecc.)
  - runSistemaTurn continua a essere chiamato uguale da /turn/end

Validazione end-to-end (zero-diff comportamentale):
- Scenario A (full HP): move + attack REGOLA_001 ✓
- Scenario B (wounded, spazio): 2x retreat REGOLA_002 ✓
- Scenario C (cornered in range, HP basso): 2x desperate attack REGOLA_001 ✓
- Scenario D (cornered fuori range, HP basso): move + attack REGOLA_001
  (approach fallback) ✓
- Bot kite 40 iter: vittoria 42 turni, P1 4/10 HP, Conquistatore(3)
  + Cacciatore(8) entrambi triggered nella stessa partita

Benefici architetturali:
1. selectAiPolicy diventa il punto unico dove aggiungere nuove regole
   (REGOLA_003 kite, stati emotivi, trait-driven overrides)
2. sistemaTurnRunner e' testabile in isolamento con mocks
3. policy.js puo' essere re-usato da logiche non-session (es. AI
   decisions in simulatori, battle previews, dashboard test)
4. Niente piu' manhattanDistance duplicata implicitamente — session.js
   ha la sua, policy.js ha la sua (entrambe trivialmente identiche,
   nessun rischio drift perche' sono pure)

Rollback: revert ripristina main post-#1357 (66f02c3b). Nessun cambio
schema, nessun cambio contract, nessun breaking change al frontend.
